+++ src/com/github/mjkuranda/retrographicsmanager/palettegenerator/PaletteGenerator.java
@@ -1,9 +1,0 @@
-package com.github.mjkuranda.retrographicsmanager.palettegenerator;
-
-import com.github.mjkuranda.retrographicsmanager.palettes.Palette;
-
-import java.awt.image.BufferedImage;
-
-public class PaletteGenerator {
-
-}